Mail namespacing correctly, registration working, composer require guzzle

+++ app/database/migrations/2015_05_15_202652_spacexstats.php
@@ -367,7 +367,9 @@ class Spacexstats extends Migration {
         }
 
         foreach($tables as $table) {
-            Schema::drop($table);
+            if ($table != 'migrations') {
+                Schema::drop($table);
+            }
         }
 	}
